Fully implement mutation
